+++ packages/desktop/src/utils/functions/capitalize.ts
@@ -1,5 +1,5 @@
 export const capitalizeFirstLetter = (string: string) => {
-  if (string.length > 0) {
+  if (string !== undefined && string.length > 0) {
     return string.charAt(0).toUpperCase() + string.slice(1)
   } else {
     return null