+++ packages/desktop/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0-alpha.3](https://github.com/TryQuiet/quiet/compare/quiet@1.1.0-alpha.2...quiet@1.1.0-alpha.3) (2023-03-13)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.1.0-alpha.2](https://github.com/TryQuiet/quiet/compare/quiet@1.1.0-alpha.1...quiet@1.1.0-alpha.2) (2023-03-09)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.1.0-alpha.1](https://github.com/TryQuiet/quiet/compare/quiet@1.1.0-alpha.0...quiet@1.1.0-alpha.1) (2023-03-09)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.1.0-alpha.0](https://github.com/TryQuiet/quiet/compare/quiet@1.0.0...quiet@1.1.0-alpha.0) (2023-03-09)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
 # [1.0.0](https://github.com/TryQuiet/quiet/compare/quiet@1.0.0-alpha.23...quiet@1.0.0) (2023-03-02)
 
 **Note:** Version bump only for package quiet